Remove kivy code from capillary and create capillarymanager to link them
